requirements.txt
grpcio-tools==1.66.1
protobuf==5.27.2
kafka-python==2.0.2